move prettyString, authorString to seperate file
