+++ src/components/tree/tipSelectedPanel.js
@@ -2,7 +2,7 @@
 /*eslint max-len: 0*/
 import React from "react";
 import {infoPanelStyles} from "../../globalStyles";
-import {prettyString, authorString} from "./treeViewFunctions";
+import {prettyString, authorString} from "../../util/stringHelpers";
 import { floatDateToMoment } from "../../util/dateHelpers";
 
 const styles = {
